+++ src/utils/upload-file-node.js
@@ -1,0 +1,13 @@
+'use strict'
+
+const { simpleFetch } = require('./simple-fetch')
+
+module.exports.uploadFileNode = (options = {}) => {
+  return simpleFetch(options.uploadUrl, {
+    method: 'PUT',
+    headers: {
+      'Content-Type': options.fileMimeType || ''
+    },
+    body: options.file
+  })
+}
